[crypto] Fix release build tests
Tests for libcrypto demand that allocation or generation of zero
length secrets result in process termination (ASSERT_DEATH in the
unittest framework).  Implemenation in the library, however, was using
a ZX_DEBUG_ASSERT instead of a ZX_ASSERT, meaning that the assert was
only turned on in a debug build, but not in a release build.

Update the implementation, promoting the assert from debug to release.
Now the parameters are always checked, and any violation of the
contract results in process termination.

Test: Build and libcrypto unit tests.
Change-Id: I69a938fa7c7acb1251b415c9ba1a5956ac22374f