[firebase_auth] Refactor cloud credentials to have its own object.
This CL introduces CloudServiceCredentials. This is a holder class that
contains all needed information to mint firebase token with a service
account.

Extracting the parsing from ServiceAccountTokenManager will allow the
Ledger to introduce new configuration file that will contaons the
service account configuration as well as the api key.

TEST=JsonSchemaTest.*, ServiceAccountTokenProviderTest.*,
     CloudServiceCredentialsTest.*

LE-554

Change-Id: I3b65076c51b011163075972d628de6bc8c105127