[kernel][object] Simplify SafeDeleter further
Before, level_ == 0 is only true for the first call to DeleteObj()
that is made when the SafeDeleter is first registered in TLS.

This means we don't need the level_ field.  The two cases of
DeleteObj() can be inlined directly into Delete().

Expand the comments in Delete().

The 'delete' loop can be simplified to a do..while loop.

Bug: none
Test: channel-test -- create_and_nest() exercises this
Change-Id: Ic490e6f5488b53a303f48674b48989348c17de55