[fidl] Ensure tools respect doc comments.
TreeVisitor is updated to visit doc comments.

The parser logic tracking the beginning / end of AST nodes is updated to
account for the fact that the beginning of certain nodes might now be an
attribute list (i.e., earlier than it was before).

A small change needed to be made to the formatter, which previously
would have indented on any newline in an interface method.  It now only
indents if that newline isn't after an attribute list.

Test: Example added to bad/goodformat.fidl
Change-Id: I73d9c5cf9d9898fc7ba0a8542f13e9116d0b83a1