[cobalt-client]: Public histogram
This is the public interface for histograms that will record data.
Users are not able to instantiate them directly, this will be done
through a peer, that will be in charge of collecting the data, and
pushing it to Cobalt.

This has two benefits, we can own the memory used for this data
structures, plus instatiation is private so updates are more isolated.

Note: Updated reverse_map so that bucket 0 lower bound is -DBL_MAX,
this we satisfy that bucket_0 = map(reverse_map(0)) which did not
happen with infinity.

TESTS=cobalt-client-tests, new_tests*150k with 0 fails.

Change-Id: Ief79dd768229ddb95c9b285b627ae5ecc55abc8b