[guest][vmm] Forbid all partially valid device ranges
DevMem is supposed to only allow adding ranges that fall between the permitted
kAddrLowerBound and kAddrUpperBound, but previously this check was done solely
on the start address, and did not consider the range having a size.

This extends AddRange to forbid the candidate range if it even partially lies
outside the permitted area.

Test: Adds to machina_unittests
Change-Id: I9ba095f72c98f16ca494d2634ab018571b884f10