[guest][docs] Change docs to use default packages
garnet/packages/default is necessary to run the guest and includes both
guest packages now, so we may as well tell users to use it instead of
individually specifying the guest packages.

Note that it's still necessary to specify the default package for the
topaz instructions because topaz only includes garnet/packages/prod/all,
which doesn't include the guests for now.

TEST=No code change

Change-Id: I2fb031237c7cbff7936a732368bd0a0470a1686f