[guest] Complete dev_mem handling for guests
The dev_mem regions were previously only passed into ZBI for ARM Zircon guests.
This adds the dev_mem regions to the E820 maps for both Linux and Zircon
guests.

It is not neccessary, nor is there any meaning, in trying to place the dev_mem
regions into the device_tree for ARM Linux guests and a comment is added to
explain this. Although we cannot pass an explicit list of device regions in
the device tree we still add a check that none of our device regions will
overlap with the physical memory that is claimed in the device tree.

Test: guest_integration_tests and launching zircon and linux guests

Change-Id: I8ad1e920b3c2363d3a61cebbf6629af5d592e7ae