[appmgr] Only publish out directories when they are mounted.
Components are responsible for serving an output directory over
PA_DIRECTORY_REQUEST, but several components do not do this. Right now
the component itself is responsible for closing the channel if it is not
needed.

We have a use case for running binaries as sandboxed components
(for tests), but the current PA_DIRECTORY_REQUEST behavior would require
modifying the binary to close the handle.

This change modifies the hub logic to only expose an "out" directory on
the filesystem when the directory is mounted, which we currently know
from USER_SIGNAL_0 sent by the Vfs ServeDirectory method.

In the future, this will be replaced with explicit events on the
channel, but for now we can use the signal behavior to solve the
problem.

CP-128: #done
TEST=$ run /boot/bin/sh -c "find /hub"; run_test_component appmgr_hub

Change-Id: I876c0f2a328fdf9092ade38999f903815766597c