[fx][boot] Add '--force-pave' option to fx boot
Certain bootserver targets are configured to be booted
with "add_fvm" set to false by default, implying the
target device should not be paved.

This toggle allows clients to override this option
with "--force-pave". This switch may be useful for targets
such as gauss and astro for experimental FVM paving support.

Change-Id: Ife45f719f60f676ec05215349ef788a08ea5666b