[wlan][brcmfmac] Implement sdio_{claim,release}_host
sdio_{claim,release}_host is a Linux DDK call which claims an SDIO
function for exclusive use.  We have a different DDK, but the brcmfmac
driver still uses this call to perform locking; hence this commit
implements a stub version of this call, which does only the
locking/unlock part.

Test: builds.

Change-Id: Ia7008508022375bc439d8564ee737049a7fdc146