[images] Write $root_build_dir/packages.json with package manifest
This manifest has JSON along the lines of:
```
{
  "available": [ "pkg1", "pkg2", ... ],
  "monolith": [ "pkg3", "pkg4", ... ]
  "preinstall": [ "pkg5", "pkg6", ... ]
  "packages": [
    {
      "build_dir": "obj/path/to/pkg1_source_dir"
      "dir": "//path/to/pkg1_source_dir"
      "name": "pkg1"
    },
    ...
  ]
}
```

This lists all the simple package names in each of the three sets
(see //build/gn/packages.gni), and then lists an entry for each package.
The `dir` and `name` together yield the GN label of the `package()`
target (i.e. `dir` contains the `BUILD.gn` defining it).  Per-package
metadata files can be found in `${build_dir}/${name}.${type}` for a
given metadata `type`.

Bug: BLD-267 #done
Test: CQ & manual inspection of packages.json output file
Change-Id: Iddda3744094123a7b6e8b8464cb30d1a8567aaba