[machina] Add VirtioComponentDevice
This class represents the in-process "configuration" piece of a device
that lives out-of-process. It manages synchronous device configuration,
and handles device interrupts that must be routed through the VMM's
interrupt controller.

A device will inherit from this class, and then provide a way to launch
the out-of-process device "component".

MAC-154

Test: N/A
Change-Id: I738b7ea7fdb4ef54895018371c631d57521df1f5