[fidlc] Refactoring: improved code tracking dependencies
Introducing a Dependencies object to properly track libray dependencies.
A specific dependency must be declared in each file constituting a
library, and use of a depedency is only allowed with an explicit using
statement.

Also introducing an AllLibraries object to track a full compilation
unit. This simplifies the code, and is a good basis to start moving code
from main.cpp into library code.

Test: no behavior change, ran unit tests, and a full build
Change-Id: Ie7a0636cb92900ebf91a7d1140466f4bd2c1e6bb