[linux_guest] Perform final mv in callers directory
Uses pushd+popd to allow the 'mv' of the final image to happen from the
callers original directory. This ensures if the -o option contained a
relative path that the path actually makes sense.

With this change the example in the README.md, which uses relative paths,
is able to work correctly.

Test: Performed steps in README.md

Change-Id: I0293190d929bd2f71f3ae2cf71add0c8e2443c86