[fidl] Fix errors with clone.h import order.
Before this change, if clone.h is imported before the fidl.h file for a
struct, ::fidl::Clone for the struct would fail because it references
the generated ::fidl::Clone that was not included yet.

This CL fixes the problem by using argument dependent lookup (ADL) for
the Clone method, but also define it in the fidl namespace as this is
where utility expects it.

TEST=Everything compiles.

Change-Id: I6d56381f55918f7baa2495d2430cb4fd84161555