[doc] Add FAQ for how to stop network connection spam on MacOS
It doesn't look like `fx setup-macos` is mentioned anywhere in the
documentation right now, and it's a very useful tool. Our current
build documentation flows don't have graceful places to add it as a one-
liner, in the meantime it'd be great to have it at least in FAQs so
it's searchable.

Change-Id: If8ca0518feac151e9f3f9811b44c0ed0124788c5