[zircon][sdio] Remove get_oob_irq proto op from sdmmc and sdio protocols
Recent changes in platform bus, allows configuration of gpio pins from child
drivers and not just the board drivers. The handling of the oob irq is moved
to the wifi driver where it belongs. This changeset removes the unnecessary
protocol from sdmmc and sdhci as it is no longer needed.

Test: fx full-build.

Change-Id: Iac2f84e83c92ff05087093e9b753e9dac56a5f5e