[debugger] Improve stepping.
Merge StepInRangeThreadController and StepIntoThreadController and use a
general StepMode enum to control how the range works. This simplifies
the integration with the StopOverThreadController and makes things more
consistent.

Uses AddressRange more consistently instead of hardcoded begin, end
pairs.

Fix a bug when stpping over that the StepThreadController only stops
on hardware breakpoints, but when used as a nested controller and
stepping out of a function, the exception is not marked as hardware. It
now uses a separate codepath that doesn't check the stop type. This was
not noticed because the test for this case didn't set the exception type
like it is in real execution (this is now fixed).

Remove special handling for "0" line numbers in the line lookup. These
need to be handled at a higher level (not done in this CL).

An additional test is added for "step into". There were no bugs for this
case but it's tricky.

TEST=unit

Change-Id: Ic4807b22ce8abf35eda7697b037c8d4f719a1c1d