[zircon][aml-sd-emmc] Add timeout for commands
sdmmc layer waits on command completions sychronously. Currently the
cmd timeout is 0, and that means the controller waits forever
for the command to complete. This changeset avoids blocking sdmmc
if a single  command is not completed within the timeout.
Most likely something else has gone wrong if the command is not
completed within this time.

Test: Booted vim2 with zircon. Tested R/W on the emmc using dd and
iotime. sdio probe is succesful. SDIO Reads and Writes are successful
With https://fuchsia-review.googlesource.com/c/garnet/+/187629/7.

Change-Id: I08dd8d82b6f3ae91544d63f5123c0620e3bbef10