[zircon][sdio]Disable ultra-high speed for now.
The wifi functionality seems to be flaky sometimes. There is a
difference in the firmware output for success and failure cases and most
likely has nothing to do with the ultra-high speed. Although there is no
direct evidence, this changeset disables ultra-high speed for now and
can be reenabled once things are stable.

Test: Build.Verified that the wifi functionality works the sameway with
this disabled.

Change-Id: Ib7c513a0d7cac65e1972fe697f0b1672acc07550