[blobfs-bench]: Avoid looping the rand_r seq
Previously for each blob, we were generating each byte from
the global seed, meaning that we consumed an element from the pseudo
random number sequence. This is O(#blobs * #bytesperblob) elements.

We reached the limit for 1MB blobs and 10k number of blobs. To avoid
hitting this, for blob we consume an element from the global sequence as
seed, and use a local pseudo random sequence to generate the entire
blob. This is O(#blobs) elements, thus the number of blobs to face this
issue is ~ 2^31.

TEST=blobfs-bench-tests

Change-Id: I46242b87af8536ee1f645407e3730c94d065df12