[auth] Define skeleton account manager impl.
AUTH-73

Note: This add a very basic account manager that adds and removes
sequential accounts locally without attempting to invoke the supplied
auth providers. The methods to get an account interface and to register
a listener are not yet functional. All auth states are hardcoded to
unknown.

Tested: Partial unit tests for the new (as yet unused) functionality.
Change-Id: I47b8f6ff7f6e69f2d6680d301e25c6b9be442871