[story_shell] Add placeholder color to module manifest
This change provides a place for module authors to specify a color
they would like to use in a placeholder that is shown between their
module being started and their module rendering pixels.

This is to facilitate the current plan of record re placeholders:
Story Shell will show a widget of specified color to finesse over
mod startup time. The fallback color is white.

Test: manual - topaz/examples/mondrian_test has a placeholder_color in
its module manifest and is read correctly. We will want an integration
test.

Change-Id: I9ba088e3dc6d67be457024bfc169e674184dd1c2