[compiler.h][fs] Remove DISABLE_THREAD_ANNOTATIONS from compiler.h
This mechanism is used by some filesystem code to conditionally
disable thread annotations. This is done for the sake of also building
code on the host.

This logic is specific to the filesystem code, and should not be part
of the public Zircon ABI. As such, this CL moves the conditionals into
the filesystem code itself.

Test: built with all compilers; CQ for mac; manually broke threading
analysis to verify that clang is still running the instrumentation.

Change-Id: I9c6453198c637fd03c1238d1c9790cb69ba11b6a