[amber] Cleanup cached connections
When configuring an update source, amber will attempt to close all cached
connections associated with the source's transport. When a source does
not contain custom transport configuration (and is therefore using Go's
default http client/transport), this attempt to close cached connections
does not succeed.

This change allows closeIdleConnections to work with both the default
http transport and configured transports.

Test=manual

Change-Id: I58a266459fe806041ef76e3deb8eb44c92f08cde