[fx] Propagate errors from amber_ctl
"fx push-package-no-publish" runs a shell script on the target to update
the requested packages, but this script is not checking errors from its
subcommands or propagating errors back to the development host.

This change checks all commands in this script, bailing out with the
first error that occurs.

PKG-235 #done

Test=manual

Change-Id: Ib66a6262bfe26131cc99befc03423329b74d1c5d