[zircon][sdio] Publish max transfer size for pio mode
Currently, the low level controller driver publishes the max transfer size depending
on whether it supports DMA. For ex: the aml-sd-emmc driver publishes to
sdio, the max_transfer_size as "4096 * 512" if the hw supports dma or "1024" if the
hardware does not support dma. With SDIO, the upper level driver can choose
whether or not it wants to use dma. In the case where dma is supported
by the host, but the upper layer does not want it, we refer to max_transfer_size of
dma mode and we end up feeding the controller with more than what it can
submit in one command. This changeset always publishes the max_transfer_size of pio
mode. When sdio sees that the upper layer does not want to use dma, it
uses the max_transfer_size of pio mode to submit the request.

Test: On vim2, Added a test function to submit a large read and saw that the right
max size was picked up. Prior to this change, the calculation of num_blocks is wrong
and hence the command fails when it reaches the controller.

ZX-2598 #Done

Change-Id: Ibd9461812b4a1c7be9c393a117780eeffa83dcd6