[linux_guest] Add depfiles to image build process
Use a depfile to always give check_image.sh a chance to run before
checking dependencies. This means we can depend on the images to get the
right thing to happen when they change but not cause errors when they
don't exist.

TEST=Build multiple times both with images and without. Observe script
called when touching the images and not called when they aren't changed.

Change-Id: Ia9d9ee4645dd7e225ee62e10a5b192e048e01196