[machina][virtio][wayland] Add WaylandDispatcher interface
This adds an OnNewConnection callback to the wayland device that
can be hooked up to a FIDL interface and provides improved testing
support.

This change also implements support for NEW_CTX, NEW_PIPE and SEND
requests. The NEW request also received a minor refactor to be
consistent with these new requests.

The device is still missing support for all response types but
this change completes the implementation of all requests.

Test: machina_unittests --gtest_filter=VirtioWlTest.*
Change-Id: Ied09da97b0f0c51973e29de79ee9597232881328