[camera] Fix fake Gralloc buffer size
In the fake implementations of sysmem, the vmo size is not
rounded up to the nearest page, although the allocated vmo
size is actually rounded up.  This was causing some cameras
to fail a buffer size check when using buffer sizes that did
not divide into evenly into the page size.

The CL adjusts the two Gralloc functions to give a buffer_size
that is rounded up to the nearest page boundry.

TEST: Ran Video_Display demo, observed camera output.
Change-Id: I309d9e3c5ff46845f23dda8915ffc391f1b57f52