[cobalt-client]: Add Collector.
Collector is the peer class in charge of collecting the
recorded data. This class uses a Logger to persist the data.

A CobaltLogger will be provided, and will be the default
for anyone instantiating the collector, through a factory
method. This abstractions simply lets us inject testing
loggers to verify what is persisted.

We require that the number of Histograms and counters
be defined upfront, to improve locality.

TEST=cobalt-client-tests

Change-Id: I78154a83e33df94e696e4cf3a24c90276ed4f01a