[ath10k] Report correct-ish RSSI and SNR
A more accurate noise floor could potentially be obtained from the
channel survey but that event comes only after we are done scanning
a channel. The original implementation uses the hard-coded default
constant, so this should be good enough for now.

NET-1352

TEST=CQ + manual 'wlan scan'. The resulting RSSI values for various APs
were very close compared to another driver we have.

Change-Id: I2a96efa880c357da384a123a49c4a7f5b2085c53