[machina] Use secondary VMO for Ethernet and introduce a copy
Changes the virtio-net host driver to allocate a new VMO to give to the
ethernet driver and packets are copied in and out of this buffer. This is
instead of having to map all of guest memory into the ethernet driver.

For simplicity the VMO we create is sized in order to guarantee that it
can never run out. Whilst this may result in slight memory wastage for
no actual performance gain, the resulting simplification is considered
worth it.

With this change we can now allow networking in the guest integration tests.

Test: Ran machina_unittests and guest_integration_tests. Launched
debian_guest and tested guest could communicate with other machines on network.

MAC-51 #done

Change-Id: I46e04d0dbf34f082af4786aa91b7ab543ac24650