[ledger] Do not return status from RunLoopFor()
RunLoopFor() is only used to ensure something didn't happen. Every usage
case EXPECTs that the loop run for the entire time. Instead of having
this return value, have RunLoopFor always run for the entire time.

The only change that this can create is that a failing test might make a
little bit more time to fail. As we do not need to optimize for failing
tests, this simplify the code.

TEST=Integration tests are still passing.

Change-Id: I71d67c0ab2bf1e6153fd4b701b35e737a65ac50b