[scenic] Adds snapshotter
Adds the ability to take a snapshot of scene graph rooted at a
branch or the whole composition. The snapshot includes attribute
and index buffers, including images. The snapshot is serialized
into flatbuffer format. The snapshot.fbs specifies the flatbuffer
schema. Uses BatchGpuUploader to read the buffres and images off
the GPU.

TEST: unittest for snapshotter
Change-Id: I6a3ff32c0f48cd94e5ef5c59bbfc995ec3dfa7ec