[media] Documentation and example updates
Based on recent feedback from sergeyu and ender, this CL adds
documentation (comments) to the audio_out.fidl file, clarifying
the usage of reference and media timestamps in the Play method
and in stream packets.

Additionally, this CL adds related comments to the usage of
Play and packet presentation timestamps in the simple_sine,
simple_sine_sync, tones and fx examples.

Test: manual (used each example, verified no functional change)
MTWN-164 #done

Change-Id: Ie2a823c193691266c3c9722067cac8d5182bc4df