[mediaplayer] add VMO info to PayloadBuffer
This CL adds the PayloadVmo class, which describes a VMO when it's
used for PayloadBuffers. It also provides a way to associate a
PayloadVmo with any number of PayloadBuffers. This code will be used
in the future for the new payload allocation scheme.

A PayloadVmo can include a FifoAllocator, which is used to implement
malloc-style allocation against the VMO.

TEST: No behavior change. CQ and manual test for regressions.
Change-Id: I50203f5c424d4ac66e3241a0bb00886d63e85400