[amber] Trigger a system update if system_image changes
We have a problem with system updates where if an update
fails partway (say we lose network connection), we won't try
to update the system because the only way we could tell if
the system needed updating was if we downloaded a new update
package, which only works once.

This patch adds an additional signal to see if the system_image
package merkle is different than what's listed in
`/boot/config/devmgr`, which tells us if any userspace system
package changed. This unfortunately won't let us know directly if
the kernel has changed, but we can infer if at least the jiri git
revision has changed for zircon, which should be sufficient for
everyone not working directly on zircon.

PKG-245 #done

TEST: manually confirmed the system update happens if
we retry updating if it fails after downloading the update
package.

Change-Id: I17d2a5a40e269e9654271cad0f5cfaad2e666619