[audio][hda] Fix an reset issue.
When resetting the HDA controller, it is important to make sure to
write a zero to the interrupt control register, instead of writing
that zero to the interrupt status register (which does nothing).
Failure to do this if there is already an unmasked interrupt pending,
while running in MSI mode, will cause you to never receive a new
interrupt (or at least, not until you manually mask and then unmask
the interrupt).

FWIW - the intel docs on this block say that a SW reset initiated
through the global control regiser will reset everything except
"certain registers".  For the record, the INTCTL register is one of
those "certain registers".

Test: Manual on Intel NUC with badly behaving zedboot which was
      leaving interrupts pending as it handed off to zircon.
Change-Id: I28fa1546204fb24161a29ec5d6e70e44e58c36b1