Revert "[module_context][story_shell] API and framework impl for mods to signal"
This reverts commit 448fab2db00a7426f5ddfbca2eb323395cf99269.

Reason for revert: Mods cannot provide a good signal for when it is ready to
display because there's no callback from scenic when it connects and draws.
Short term solution is to have a timeout (estimated from cobalt data) on the
story shell instead. Long term solution is for scenic to provide signals via
views.

Original change's description:
> [module_context][story_shell] API and framework impl for mods to signal
> that the view is ready, and for story shell to receive this signal.
>
> How story shell will communicate to user shell is yet to be determined
> so I will put this out in another CL once it's decided.
>
> MI4-1341 #done
>
> TESTED: paved, integration tests
> Change-Id: I844b312cfc634d5e1177bb633de4b67e77a0a22f

TBR=vardhan@google.com,mesch@google.com,djmurphy@google.com,lindkvist@google.com,alexmin@google.com

# Not skipping CQ checks because original CL landed > 1 day ago.

Change-Id: I58e758f8c626348a23925995c6487b9a6a2d502e