[mediaplayer] add VmoPayloadAllocator
This CL adds VmoPayloadAllocator, a class that allocates payload
buffers from one or more VMOs.  It also adds two interfaces that
are used to expose features of the VMO allocator: PayloadVmos and
ExternalPayloadVmos. These definitions will be used in the new
payload allocation scheme, but are currently unused.

LocalMemoryPayloadAllocator has also been updated to make a call
that was commented out previously.

TEST: No behavior change. CQ tests for regressions.

Change-Id: Ifc2aeb9ea64c6ae42e5b1c5aec8196f2b484d7b4