[media] Make persisted property name consistent
As part 3 of making the audio subsystem consistent in its usage
of 'gain_db' instead of a mixture of 'db_gain', 'gain_db', this
CL changes the persisted setting for device volume from db_gain
to gain_db.

Test: manually verified scenarios with: no config file; only the
old property; only the new property; and (for defense in depth)
the seemingly-impossible scenarios of config files with neither
property, and with both properties.

Change-Id: I73749e461c022be57474e42a1939ac571a0edbba