[audio][frank] Add a fake product configuration package for 'Frank'
Add a package which carries default product specific configuration
data for the Frank target.

Note:  This is not the long term solution.  The package folks are in
the process of devising better ways of mixing in customizations to
packages based of various criteria.  Eventually it should be possible
to have this type of platform or product specific configuration be
expressed as optional components of the audio_core package.  Until
that is possible, however, we create a specific package for a specific
product which needs to be selected by product specific build files.

Test: Manual testing on Frank HW.  Paved and confirmed that the
      configuration file was in the proper location, and that the file
      was properly chosen as the initial default.
Change-Id: Ib9231d527afd30ee3623cf0c1144d94ccd55bd98