[kernel][socket] Threshold based signalling on sockets.
Adds 2 new socket object properties - read and write thresholds,
and 2 new signals ZX_SOCKET_WRITE_THRESHOLD/ZX_SOCKET_READ_THRESHOLD.
We signal the read when we have more bytes than the read threshold
that can be read, and we signal the write when there is more space
than the write threshold available for writing. We can wait for the
new signals on zx_object_wait_one/many() instead of busy waiting
until enough data is available to read or enough space is available
for writing.

ZX-2200.

Test: Added a unit test to core/socket-test to make sure the signals
are being asserted and de-asserted as data gets read and written.

Change-Id: I1214df29570435ec73e716167bfa9206c8ffcdcb