[add_mod_call] Don't call UpdateMod
The add mod call used to first call an update mod, and then call an add
only if the update failed due to the mod not being found. This does not
work well with the new intent handling, and there is already logic in
story_controller which will make sure we don't restart modules
unnecessarily.

TESTED=add_mod_command_runner_unittest and manually.

Change-Id: I34f1654dc0a09e97e87253c18e4983bb66e07036