[services] Fix Lookup and Readdir on /svc
This correctly implements Lookup and Readdir for ServiceProviderDirImpl,
which is now possible thanks to the new CreateNestedEnvironment API.

The (incorrect, but required for backwards compatibility) behavior for
backing_dir is still supported, and will be removed when we remove
backing_dir.

CP-124 #comment

TESTED=service_provider_dir_unittests, and manually tested topaz.
       Integration tests coming soon

Change-Id: Ie1ef3662b503c92e0058d4d75ca427e56670fbd9