[scenic] Snapshot API
Adds 'requestSnapshotHACK' method to ViewContainer. This method
is accessible from ChildView Dart class. It returns the snapshot
data as a VMO buffer. This buffer can then be handed to
SnapshotViewer widget, which loads and displays the snapshot.

This API is Experimental in nature. Hence the word HACK appended
to its name. The API design is in flux and will change.

- Includes 'snapshot' scenic app that displays the snapshot in
  a view.

TEST: Manual

Change-Id: Ibca7c1d141f0d094fd934d6b2db6e8a5da71b6ae