[zircon][sdio]Perform tuning when setting ultra high speed
Currently on VIM2, at times, the wifi chipset does not run all the way.
This is mainly being attributed to power issues. Although not related,
switching to ultra-high speed in sdio is disabled as a safety measure
until the wifi stack is more stable. This changeset adds tuning
procedure when switching to ultra-high speed. Tuning procedure helps
determine any read delays required and the right frequency when the
ultra-high speed is re-enabled.

Test: On vim2, enabled ultra high speed and verified the wifi functionality.

Change-Id: I32f0dbafbcdde32c7d220e38a1c43fc57e492284