[user_intelligence_provider] Enqueue service connections during restart
It is known that in the event of an agent crash, its services will go down
well in advance of the agent controller. Agent restart must wait for the agent
controller for the restart attempt to succeed, so we must queue incoming
service requests during this time to prevent thrashing and spurious give-up.

MI4-1353 #comment

Test: hard-coded agent to crash on query
      Before: included log line [WARNING:...] No speech-to-text agent loaded;
              multiple restart attempts
      After: no such line; single restart attempt

      hard-coded agent to crash on auth; verified gave up; no unexpected
      crashes

Change-Id: I276a22e82a69025e47d3e34023023ec4cbdd1673