[guest] Unify bootloader device configuration
Defines a new base class PlatformDevice that contains methods for a device to construct
device tree and ZBI information. Adds this to the existing platform devices and moves
the configuration code from the guest launch code to the device.

With a common base class we also collect the devices instantiated in the vmm main and
pass them setup_zircon and setup_linux. This removes the duplication of those setup
functions from knowing which devices were instantiated.

Test: runtests guest_integration_test. Launch zircon and linux guests manually

Change-Id: Ic455c6e99fca3b6e2a8d6cc338180d18bb926943