[guest] Generate RAM around device memory ranges
This removes the assumption that there will be a single RAM range from
`0..cfg.memory()`, and instead generates RAM regions around any
potential device memory.

Currently, by virtue of anything in dev_mem being defined as not
overlapping with the `0..cfg.memory()` range, this commit is ultimately
a no-op and still results in a single region being generated. This
does allow for eventually expanding what is in dev_mem to include
devices that are currently hard coded (and might overlap with the RAM
range) to be treated uniformly and reliably not overlap.

Test: fx test guest_integration_test

Change-Id: I0ab1671f2740798e332110fd2c32e9b2c2484306