[mediaplayer] fix demux threading issue
The ffmpeg demux notifies its client when it completes initialization
and is ready to stream packets. This notification starts in a worker
thread and calls out to thread-unsafe code that expects to be called
on the player's main thread. This CL changes the notification so it
posts to the correct thread.

TEST: no new behavior, fix manually verified on VIM
Change-Id: Ibffa88a436accaaf11923befae42d94298119681