[media][audio-core] Fix leaking AudioOut instances.
Fix 2 causes of AudioOut instances being leaked.

The first is that the AudioOut was failing to remove itself from the
set of active AudioOuts in the AudioDeviceManager during shutdown.
Casual scanning of the change history suggests that this was
introduced during the v2 migration of AudioRenderer.

The second was that the AudioOut was failing to remove it's dedicated
link to the ThrottleOutput.  This ended up causing a circular
reference between the link to the throttle output and the AudioOut
instance itself.

Moving forward; finally getting around to fixing MTWN-52 in the mixer
is probably the best was to make the code more robust re: the second
failure.  Instead of getting rid if the throttle output entirely, it
is probably best to just make it less of a special case.
Specifically, any time an AudioOut's destination link count hits zero,
it should become linked to the throttle output.  Any time a new Output
is added to an AudioOut's destination set, if the AudioOut is linked
_only_ to the throttle out, it should have the the throttle output
removed.

Testing:
Repro for this is simple.  One can simply boot, and run...
fx shell "while true; do signal_generator --dur=0.1; done"
and watch the memory stats rise.

During investigation, manual instrumentation was added to track the
count of each active object in the system, and to periodically dump
the counts.  With this instrumentation in place, it was easy to see
that AudioOut and AudioLink instances were not going away.  Once the
causes were tracked down, instrumentation showed the counts remaining
where they should instead of growing without bound, and ps showed
committed memory remaining stable at ~2MB in a debug build.

Test: Manual; see above for details.
MTWN-165 #done

Change-Id: Id06ba38f0095a033c058794d8d152202e9c63009