[escher] Lazily Create Frame for BatchGpuUploader
Don't instantiate a frame for the BatchGpuUploader (in effect
instantiating a commandbuffer) until a client requests a Reader or
Writer that will use the backing commandbuffer. This reduces overhead of
creating a commandbuffer that is unused, which submits empty begin/end
commands to the GPU, and causes unnecessary driver churn.

Tested:
- fx shell run escher_tests: unittests still pass
- took traces of 1) app that continuously uploads gpu commands and of 2)
an app that initializes with gpu uploads and then reuses scene geometry.
Verified that behavior of the BatchGpuUploader is as expected in both
scenarios.

SCN-951 #done
ES-135 #done

Change-Id: I2c5f46fd82415147fe17e6afb96290c9c701af4e