[appmgr/fuchsia.sys] Create an event for output directory mounting.
We currently do not have an indication that a component is ready,
instead we have to rely on the output directory showing up using the
Glob interface.

This change adds an OnDirectoryReady event that triggers whenever a
component has its output directory mounted in hub. While this is
equivalent to checking the hub output directory, it is a cleaner
solution that does not depend on hub access.

This can help us deal with race conditions between component starting
and observing state in integration tests.

CP-130: #done
TEST=appmgr_unittests

Change-Id: I11a1afd6c83e2bda4fe6e9d2c4c294bc94acee36