[debugger] Add sym-info command.
Currently this justs list information for local variables. I will use it
for debugging some expression bugs, and it may be useful in the future.

The EvalContext is expanded to support variable lookup that returns the
Symbol object rather than the value it refers to. This has a nice side
effect of cleaning up the the existing implementation a bit.

TEST=manual

Change-Id: I8e290cddb1be8b913b31ffcd66f795435820f740