[sysmem] Introduce basic functionality to sysmem
With this CL,  Sysmem can now be used to allocate
basic bufffer collections through the AllocateCollection
function.

BufferUsages are currently ignored, and multiplanar formats
are not supported.  This is really just to test out the basics
of Sysmem.

A client library with unit tests is broken out into the next CL.

TEST = CQ

Change-Id: I851c71e41373a577eab59f7872c25856ae409cb9