[debugger] Fix printing values.
Sets the "value" flag on the DwarfExprEval when executing DW_OP_reg? to
indicate that the value of the register is a value rather than a
pointer.

Adds the expected result type to all DwarfExprEval tests.

Fix a crash evaluating an expression if the success callback deletes the
DwarfExprEval class. This was already handled correctly in the error
case but the success case was not updated.

TEST=unit, manual

Change-Id: I7820247788abd245bcf3029945c326ba9feeb79a