[guest] Use DevMem for all device ranges
Retrieves all regions of memory that have a trap mapping and adds those
as device regions to dev_mem. By adding all the actual device regions
hardcoding of device ranges, that just hope to contain all devices, in
the bootloader configuration can be removed. Nothing needs to be added
as dev_mem regions were already being added to bootloader configuration.

To allow for adding all devices to dev_mem the range restriction needs
to be removed from DevMem. It is safe to do this since it is now
separately validated that RAM will not be assigned to overlap with
any range in DevMem.

With DevMem no longer needing a `kAddrLowerBound` constant the vmm
main now calls this same value `kFirstDynamicDeviceAdddr` and performs
simple allocation starting from there.

Test: fx test guest_integration_test

Change-Id: I1453aa788eaf76ea52df27041e73fbb688188251