[scenic] ObjectLinker enhancements
Using template sugar, allow the ObjectLinker to be used with handles of
any peerable type.  SFINAE is used to ensure that trying to use a
non-peerable handle results in a compiler error.

Ensure links correctly mark themselves as de-initialzed when the peer on
the other end is destroyed.

Links to be replaced via move operations, as well as to be constructed in an
invalid state.  This helps clients which need to "transfer" Links
between things from time to time.

SCN-829 #done
SCN-956 #done
SCN-914 #comment
Test: object_linker_unittest

Change-Id: Iaf4807463f69e1c824561a181c183aabfca80be6